refactor(dev-core): move resolveCurrentProject to lib/workspace
Suggestion #6 from PR #100 review. Colocates the function with
resolveRepoFromCwd (which it calls), removes the command-module
import side-effect surface from tests, and generalizes the signature
to `<P extends { repo, localPath? }>` so any project shape with those
fields can be resolved.

No behavior change. Test imports simplified to a single source.
